# lib/api-client-react/src/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEPZIA Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  condition: string;
  category: string;
  location: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  images?: string[];
  status: string;
  featured?: boolean;
  boosted?: boolean;
  viewCount?: number;
  sellerId: string;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerAvatar?: string | null;
  sellerVerified?: boolean;
  /** @nullable */
  sellerRating?: number | null;
  /** @nullable */
  contactPhone?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ListingsPage {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
}

export interface ListingInput {
  title: string;
  description?: string;
  price: number;
  condition: string;
  category: string;
  location: string;
  latitude?: number;
  longitude?: number;
  images?: string[];
  contactPhone?: string;
  status?: string;
  featured?: boolean;
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  condition?: string;
  category?: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  images?: string[];
  contactPhone?: string;
  status?: string;
  featured?: boolean;
}

export interface ListingStatusUpdate {
  status: string;
}

export interface ReportInput {
  reason: string;
  details?: string;
}

export interface OfferInput {
  amount: number;
  message?: string;
}

export type OfferResponseAction = typeof OfferResponseAction[keyof typeof OfferResponseAction];


export const OfferResponseAction = {
  accepted: 'accepted',
  rejected: 'rejected',
  countered: 'countered',
} as const;

export interface OfferResponse {
  action: OfferResponseAction;
  counterAmount?: number;
}

export interface Offer {
  id: number;
  listingId: number;
  /** @nullable */
  listingTitle?: string | null;
  /** @nullable */
  listingImage?: string | null;
  buyerId: string;
  /** @nullable */
  buyerName?: string | null;
  sellerId: string;
  amount: number;
  /** @nullable */
  message?: string | null;
  status: string;
  /** @nullable */
  counterAmount?: number | null;
  createdAt: string;
}

export interface AppNotification {
  id: number;
  userId: string;
  type: string;
  title: string;
  body: string;
  /** @nullable */
  data?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  count?: number;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  isVerified?: boolean;
  isAdmin?: boolean;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  bio?: string;
  phone?: string;
  location?: string;
}

export interface SellerProfile {
  id: number;
  clerkId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  isVerified?: boolean;
  listingCount: number;
  rating: number;
  createdAt: string;
}

export interface Conversation {
  id: number;
  listingId: number;
  /** @nullable */
  listingTitle?: string | null;
  /** @nullable */
  listingImage?: string | null;
  buyerId: string;
  sellerId: string;
  /** @nullable */
  otherPartyName?: string | null;
  /** @nullable */
  otherPartyAvatar?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount?: number;
  createdAt: string;
}

export interface ConversationInput {
  listingId: number;
  sellerId: string;
  initialMessage?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface MarketplaceStats {
  totalListings: number;
  activeListings: number;
  totalUsers: number;
  totalCategories: number;
  totalSellers: number;
  listingsToday?: number;
}

export interface CategoryStat {
  category: string;
  count: number;
  slug?: string;
}

export type ListListingsParams = {
q?: string;
category?: string;
location?: string;
condition?: string;
minPrice?: number;
maxPrice?: number;
sort?: string;
featured?: boolean;
page?: number;
limit?: number;
};

export type GetFeaturedListingsParams = {
limit?: number;
};

export type GetRecentListingsParams = {
limit?: number;
};

export type GetNearbyListingsParams = {
lat: number;
lng: number;
/**
 * Radius in km (default 50)
 */
radius?: number;
limit?: number;
};

export type AdminListListingsParams = {
status?: string;
};