# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NEPZIA Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get public seller profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
  /users/{id}/listings:
    get:
      operationId: getSellerListings
      tags: [users]
      summary: Get listings by seller
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Seller's listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"

  # ─── Notifications ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Listing:
      type: object
      required: [id, title, price, condition, category, location, status, createdAt, sellerId]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        location: { type: string }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        status: { type: string }
        featured: { type: boolean }
        boosted: { type: boolean }
        viewCount: { type: integer }
        sellerId: { type: string }
        sellerName: { type: ["string", "null"] }
        sellerAvatar: { type: ["string", "null"] }
        sellerVerified: { type: boolean }
        sellerRating: { type: ["number", "null"] }
        contactPhone: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
...
    UserProfile:
      type: object
      required: [id, clerkId, createdAt]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        name: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        avatar: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isAdmin: { type: boolean }
        createdAt: { type: string }
    UserProfileUpdate:
      type: object
      properties:
        name: { type: string }
        bio: { type: string }
        phone: { type: string }
        location: { type: string }
    SellerProfile:
      type: object
      required: [id, clerkId, createdAt, listingCount, rating]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        name: { type: ["string", "null"] }
        avatar: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        isVerified: { type: boolean }
        listingCount: { type: integer }
        rating: { type: number }
        createdAt: { type: string }
...