# lib/api-zod/src/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEPZIA Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  clerkId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  isVerified?: boolean;
  isAdmin?: boolean;
  createdAt: string;
}